Remove "run or adapt the code"
I removed this part from the installation guide. I think it is sufficient to know how to run the code. To adapt the code, you really need more GithHub skills then those listed in the installation guide. Maybe we could add a "report an issue" section, to indicate that comments are always welcome.